+++ packages/api-explorer/src/utils/index.ts
@@ -27,8 +27,4 @@ export { highlightHTML } from './highlight'
 export { buildMethodPath, buildTypePath, diffPath, oAuthPath } from './path'
 export { getLoded } from './lodeUtils'
 export { useWindowSize } from './useWindowSize'
-export {
-  IApixEnvAdaptor,
-  StandaloneEnvAdaptor,
-  EnvAdaptorConstants,
-} from './envAdaptor'
+export * from './envAdaptor'