+++ packages/extension-api-explorer/src/utils.spec.ts
@@ -23,31 +23,29 @@
  SOFTWARE.
 
  */
-import React, { FC } from 'react'
 import {
-  ComponentsProvider,
-  Flex,
-  FlexItem,
-  Heading,
-  Spinner,
-} from '@looker/components'
-import { ThemeOverrides } from '@looker/api-explorer/src/utils'
+  getThemeOverrides,
+  ThemeOverrides,
+} from '@looker/api-explorer/src/utils'
+import { ExtensionSDK, LookerHostData } from '@looker/extension-sdk'
+import { ExtensionEnvAdaptor } from './utils'
 
-export interface LoaderProps {
-  themeOverrides: ThemeOverrides
-}
-
-export const Loader: FC<LoaderProps> = ({ themeOverrides }) => (
-  <ComponentsProvider {...themeOverrides}>
-    <Flex flexDirection="column" justifyContent="center" mt="25%">
-      <FlexItem alignSelf="center">
-        <Spinner color="key" size={150} />
-      </FlexItem>
-      <FlexItem mt="large" alignSelf="center">
-        <Heading color="key" as="h2">
-          Loading API Specifications
-        </Heading>
-      </FlexItem>
-    </Flex>
-  </ComponentsProvider>
-)
+describe('ExtensionEnvAdaptor', () => {
+  test.each([
+    [undefined, getThemeOverrides(false)],
+    ['standard', getThemeOverrides(true)],
+    ['embed', getThemeOverrides(false)],
+    ['spartan', getThemeOverrides(false)],
+  ])(
+    'returns correct font overrides',
+    (hostType?: string, expectedOverrides?: ThemeOverrides) => {
+      expect(
+        new ExtensionEnvAdaptor({
+          lookerHostData: {
+            hostType,
+          } as Readonly<LookerHostData>,
+        } as ExtensionSDK).themeOverrides()
+      ).toEqual(expectedOverrides)
+    }
+  )
+})