+++ packages/run-it/src/RunItProvider.tsx
@@ -25,12 +25,12 @@
  */
 
 import React, { createContext, FC, ReactElement } from 'react'
-import { Looker40SDK, Looker31SDK } from '@looker/sdk'
+import { IAPIMethods } from '@looker/sdk-rtl'
 import { RunItConfigurator } from './components'
 
 export interface RunItContextProps {
   basePath: string
-  sdk?: Looker40SDK | Looker31SDK | undefined
+  sdk?: IAPIMethods | undefined
   configurator: RunItConfigurator
 }