+++ packages/run-it/src/components/ConfigForm/index.ts
@@ -24,13 +24,6 @@
 
  */
 
-export { ConfigDialog } from './ConfigDialog'
-export { ConfigForm } from './ConfigForm'
-export {
-  RunItConfigurator,
-  defaultConfigurator,
-  validateUrl,
-  validLocation,
-  RunItConfigKey,
-  RunItValuesKey,
-} from './configUtils'
+export * from './ConfigDialog'
+export * from './ConfigForm'
+export * from './configUtils'