+++ packages/sdk-rtl/src/apiMethods.ts
@@ -37,7 +37,7 @@ import {
 import { IAuthSession } from './authSession'
 
 /**
- * Creates an "sdk" to be used with the Typescript SDK's funcs.ts file
+ * Creates an "sdk" to be used with the TypeScript SDK's funcs.ts file
  * @param authSession authentication session
  * @param apiVersion version of API to use (e.g. "3.1" or "4.0")
  * @param sdkVersion Looker release version (e.g. "21.4")