+++ packages/api-explorer/CHANGELOG.md
@@ -5,6 +5,35 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### [0.9.6](https://www.github.com/looker-open-source/sdk-codegen/compare/api-explorer-v0.9.5...api-explorer-v0.9.6) (2021-04-27)
+
+
+### Features
+
+* APIX now uses google font appropriately ([#636](https://www.github.com/looker-open-source/sdk-codegen/issues/636)) ([ed932d6](https://www.github.com/looker-open-source/sdk-codegen/commit/ed932d6744fd3c102bfff7417643667bb57aee00))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @looker/run-it bumped from ^0.9.5 to ^0.9.6
+
+### [0.9.5](https://www.github.com/looker-open-source/sdk-codegen/compare/api-explorer-v0.9.4...api-explorer-v0.9.5) (2021-04-27)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @looker/run-it bumped from ^0.9.4 to ^0.9.5
+    * @looker/sdk bumped from ^21.4.2 to ^21.4.3
+    * @looker/sdk-codegen bumped from ^21.0.14 to ^21.0.15
+    * @looker/code-editor bumped from ^0.0.1 to ^0.0.2
+  * devDependencies
+    * @looker/sdk-codegen-scripts bumped from ^21.0.14 to ^21.0.15
+    * @looker/sdk-node bumped from ^21.4.3 to ^21.4.4
+
 ### [0.9.4](https://www.github.com/looker-open-source/sdk-codegen/compare/api-explorer-v0.9.3...api-explorer-v0.9.4) (2021-04-26)