+++ packages/code-editor/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## [0.1.0](https://www.github.com/looker-open-source/sdk-codegen/compare/code-editor-v0.0.2...code-editor-v0.1.0) (2021-04-27)
+
+
+### Features
+
+* APIX now uses google font appropriately ([#636](https://www.github.com/looker-open-source/sdk-codegen/issues/636)) ([ed932d6](https://www.github.com/looker-open-source/sdk-codegen/commit/ed932d6744fd3c102bfff7417643667bb57aee00))
+
+### [0.0.2](https://www.github.com/looker-open-source/sdk-codegen/compare/code-editor-v0.0.1...code-editor-v0.0.2) (2021-04-27)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @looker/sdk-codegen bumped from ^21.0.14 to ^21.0.15
+
 ### 0.0.1 (2021-04-26)