+++ packages/sdk-node/CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### [21.4.4](https://www.github.com/looker-open-source/sdk-codegen/compare/sdk-node-v21.4.3...sdk-node-v21.4.4) (2021-04-27)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @looker/sdk bumped from ^21.4.2 to ^21.4.3
+
 ### [21.4.3](https://www.github.com/looker-open-source/sdk-codegen/compare/sdk-node-v21.4.2...sdk-node-v21.4.3) (2021-04-26)
 
 
@@ -52,4 +61,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [21.0.4]
 
-- introduces `@looker/node-sdk` for using the Looker Typescript SDK with Node
+- introduces `@looker/node-sdk` for using the Looker TypeScript SDK with Node